refactor: rename `CredentialListEntry`
