+++ src/lib/components/CredentialListItem.svelte
