+++ src/lib/connections/ConnectionData.svelte
@@ -2,7 +2,7 @@
   import EnvelopeSimple from '~icons/ph/envelope-simple-light';
   import House from '~icons/ph/house-light';
 
-  import CredentialListEntry from '../components/CredentialListEntry.svelte';
+  import CredentialListItem from '../components/CredentialListItem.svelte';
 
   const data = [];
 </script>
@@ -16,12 +16,12 @@
     <div class="text-center text-[22px]/[30px] font-semibold text-slate-700 dark:text-grey">Data shared</div>
   </div>
   <div class="flex w-full flex-col rounded-2xl border border-slate-200 bg-white p-1 dark:border-slate-600 dark:bg-dark">
-    <CredentialListEntry title="Email address" description="ferris.rustacean@example.org">
+    <CredentialListItem title="Email address" description="ferris.rustacean@example.org">
       <!-- <span slot="icon"><EnvelopeSimple class="h-6 w-6" /></span> -->
-    </CredentialListEntry>
-    <CredentialListEntry title="Home address" description="Sunset Rd 1337, Iron Islands">
+    </CredentialListItem>
+    <CredentialListItem title="Home address" description="Sunset Rd 1337, Iron Islands">
       <!-- <span slot="icon"><House class="h-6 w-6" /></span> -->
-    </CredentialListEntry>
+    </CredentialListItem>
   </div>
 {:else}
   <div class="flex h-full flex-col items-center justify-center">
